refactor: adjust types to live in chat folder
