+++ src/ChatApp/Webs/chatapp/src/app/chat/(components)/types.ts
@@ -6,10 +6,3 @@ export interface Conversation {
   avatar?: string;
   selected?: boolean;
 }
-
-export interface Message {
-  id: number;
-  sender: string;
-  content: string;
-  time: string;
-}
